test: clean tmp and provide benchmark file and code comments
